fix(org): drop redundant u64::from(nlink()) — Linux production clippy
MetadataExt::nlink() already returns u64 on every Unix, so u64::from(...) is a
useless_conversion that fails Linux `cargo clippy --lib -D warnings`. Use the
value directly; the Windows arm still converts its u32 nNumberOfLinks.

+++ net/crates/net/src/adapter/net/behavior/org_revocation.rs
@@ -1718,7 +1718,8 @@ pub(crate) fn lock_state_file(path: &Path) -> Result<std::fs::File, OrgRevocatio
     #[cfg(unix)]
     let nlink = {
         use std::os::unix::fs::MetadataExt;
-        u64::from(lock.metadata().map_err(io)?.nlink())
+        // `MetadataExt::nlink()` is already `u64` on every Unix — no conversion.
+        lock.metadata().map_err(io)?.nlink()
     };
     #[cfg(windows)]
     let nlink = {
